requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.9.2
psycopg[binary]==3.2.3
sqlalchemy==2.0.36
python-dotenv==1.0.0
numpy==1.26.4
